+++ enerzyme/models/activation.py
@@ -5,11 +5,31 @@ import torch.nn.functional as F
 from torch import Tensor, nn
 
 
-def shifted_softplus(x: Tensor) -> Tensor:
-    return 
+class BaseScaledTemperedActivation(nn.Module):
+    def __init__(self, dim_feature: int=1, initial_alpha: float=1.0, initial_beta: float=1.0, learnable: bool=False) -> None:
+        super().__init__()
+        if float(initial_alpha) == 1.0 and float(initial_beta) == 1.0 and not learnable:
+            self.simple = True
+            self.alpha = 1.0
+            self.beta = 1.0
+            self._activation_fn = self.simple_activation_fn
+        else:
+            self.simple = False
+            self.register_parameter("alpha", nn.Parameter(torch.full((dim_feature,), initial_alpha), requires_grad=learnable))
+            self.register_parameter("beta", nn.Parameter(torch.full((dim_feature,), initial_beta), requires_grad=learnable))
+            self._activation_fn = self.activation_fn
+
+    def simple_activation_fn(self, x: Tensor) -> Tensor:
+        ...
+
+    def activation_fn(self, x: Tensor) -> Tensor:
+        ...
+
+    def forward(self, x: Tensor) -> Tensor:
+        return self._activation_fn(x)
 
 
-class ShiftedSoftplus(nn.Module):
+class ShiftedSoftplus(BaseScaledTemperedActivation):
     def __init__(self, dim_feature: int=1, initial_alpha: float=1.0, initial_beta: float=1.0, learnable: bool=False) -> None:
         """
         Shifted softplus activation function with learnable feature-wise parameters:
@@ -26,34 +46,69 @@ class ShiftedSoftplus(nn.Module):
             initial_beta (float):
                 Initial "temperature" beta of the softplus function.
         """
-        super().__init__()
+        super().__init__(dim_feature, initial_alpha, initial_beta, learnable)
         self.log2 = math.log(2.0)
-        if float(initial_alpha) == 1.0 and float(initial_beta) == 1.0 and not learnable:
-            self._shifted_softplus = self.simple_shifted_softplus
-            self.alpha = 1.0
-            self.beta = 1.0
-        else:
-            self.register_parameter("alpha", nn.Parameter(torch.full((dim_feature,), initial_alpha), requires_grad=learnable))
-            self.register_parameter("beta", nn.Parameter(torch.full((dim_feature,), initial_beta), requires_grad=learnable))
-            self._shifted_softplus = self.scaled_shifted_softplus
     
-    def simple_shifted_softplus(self, x: Tensor) -> Tensor:
+    def simple_activation_fn(self, x: Tensor) -> Tensor:
         return F.softplus(x) - self.log2
     
-    def scaled_shifted_softplus(self, x: Tensor) -> Tensor:
+    def activation_fn(self, x: Tensor) -> Tensor:
         return self.alpha * torch.where(
             self.beta != 0,
             (F.softplus(self.beta * x) - self.log2) / self.beta,
             0.5 * x
         )
 
-    def forward(self, x: Tensor) -> Tensor:
-        return self._shifted_softplus(x)
+
+class Swish(BaseScaledTemperedActivation):
+    """
+    Swish activation function with learnable feature-wise parameters:
+    f(x) = alpha*x * sigmoid(beta*x)
+    sigmoid(x) = 1/(1 + exp(-x))
+    For beta -> 0  : f(x) -> 0.5*alpha*x
+    For beta -> inf: f(x) -> max(0, alpha*x)
+
+    Arguments:
+        num_features (int):
+            Dimensions of feature space.
+        initial_alpha (float):
+            Initial "scale" alpha of the "linear component".
+        initial_beta (float):
+            Initial "temperature" of the "sigmoid component". The default value
+            of 1.702 has the effect of initializing swish to an approximation
+            of the Gaussian Error Linear Unit (GELU) activation function from
+            Hendrycks, Dan, and Gimpel, Kevin. "Gaussian error linear units
+            (GELUs)."
+    """
+
+    def __init__(
+        self, dim_feature: int, initial_alpha: float=1.0, initial_beta: float=1.702, learnable: bool=True
+    ) -> None:
+        """ Initializes the Swish class. """
+        super().__init__(dim_feature, initial_alpha, initial_beta, learnable)
+
+    def simple_activation_fn(self, x: Tensor) -> Tensor:
+        return F.silu(x)
+
+    def activation_fn(self, x: Tensor) -> Tensor:
+        """
+        Evaluate activation function given the input features x.
+        num_features: Dimensions of feature space.
+
+        Arguments:
+            x (FloatTensor [:, num_features]):
+                Input features.
+
+        Returns:
+            y (FloatTensor [:, num_features]):
+                Activated features.
+        """
+        return self.alpha * x * torch.sigmoid(self.beta * x)
 
 
 ACTIVATION_REGISTER = {
     "shifted_softplus": ShiftedSoftplus,
-    "swish": ...,
+    "swish": Swish,
 }
